requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
django-guardian==2.4.0
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.3.0
importlib-metadata==4.8.2
Markdown==3.3.6
phonenumbers==8.12.39
pre-commit==2.17.0
psycopg2==2.9.2
pytz==2021.3
PyYAML==5.4.1
sqlparse==0.4.2
tornado==6.1
watchdog==2.1.6
websocket-client==1.2.3
zipp==3.6.0
